fix handleBatch function signature

+++ src/sqs-consumer.ts
@@ -16,7 +16,7 @@ export interface SqsConsumerOptions {
     sqsEndpointUrl?: string;
     s3EndpointUrl?: string;
     handleMessage?(message: SqsMessage): Promise<void>;
-    handleBatch?(messages: SqsMessage[]): Promise<Message[]|undefined>;
+    handleBatch?(messages: SqsMessage[]): Promise<Message[]|void>;
     parsePayload?(payload: any): any;
     transformMessageBody?(messageBody: any): any;
     // Opt-in to enable compatibility with
@@ -60,7 +60,7 @@ export class SqsConsumer {
     private events = new EventEmitter();
     private connErrorTimeout = 10000;
     private handleMessage?: (message: SqsMessage) => Promise<void>;
-    private handleBatch?: (messagesWithPayload: SqsMessage[]) => Promise<Message[]|undefined>;
+    private handleBatch?: (messagesWithPayload: SqsMessage[]) => Promise<Message[]|void>;
     private parsePayload?: (payload: any) => any;
     private transformMessageBody?: (messageBody: any) => any;
     private extendedLibraryCompatibility: boolean;
@@ -172,7 +172,7 @@ export class SqsConsumer {
             }));
 
             const messagesToDelete = await this.handleBatch(messagesWithPayload);
-            if (messagesToDelete?.length) 
+            if (messagesToDelete && messagesToDelete?.length) 
                 await this.deleteBatch(messagesToDelete);
             else if (messagesToDelete === undefined)
                 await this.deleteBatch(messages);
